+++ docker-compose.yml
@@ -2,6 +2,8 @@ version: '3'
 services:
    my-mongoDB:
       image: mongo:latest
+      ports:
+         - 27017:27017
       volumes:
         - db-data:/data/db
         - mongo-config:/data/configdb
